Move skin editor to overlays namespace
