+++ osu.Game.Tests/Visual/Gameplay/TestSceneSkinEditor.cs
@@ -9,11 +9,11 @@ using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Testing;
 using osu.Game.Overlays;
 using osu.Game.Overlays.Settings;
+using osu.Game.Overlays.SkinEditor;
 using osu.Game.Rulesets;
 using osu.Game.Rulesets.Osu;
 using osu.Game.Screens.Play.HUD.HitErrorMeters;
 using osu.Game.Skinning;
-using osu.Game.Skinning.Editor;
 using osuTK.Input;
 
 namespace osu.Game.Tests.Visual.Gameplay
