+++ osu.Game.Tests/Visual/Gameplay/TestSceneSkinEditorComponentsList.cs
@@ -7,9 +7,9 @@ using NUnit.Framework;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Game.Overlays;
+using osu.Game.Overlays.SkinEditor;
 using osu.Game.Rulesets;
 using osu.Game.Rulesets.Osu;
-using osu.Game.Skinning.Editor;
 
 namespace osu.Game.Tests.Visual.Gameplay
 {
