+++ osu.Game.Tests/Visual/Navigation/TestSceneEditDefaultSkin.cs
@@ -8,8 +8,8 @@ using NUnit.Framework;
 using osu.Framework.Allocation;
 using osu.Framework.Testing;
 using osu.Game.Overlays.Settings.Sections;
+using osu.Game.Overlays.SkinEditor;
 using osu.Game.Skinning;
-using osu.Game.Skinning.Editor;
 
 namespace osu.Game.Tests.Visual.Navigation
 {
