+++ osu.Game/Overlays/SkinEditor/SkinBlueprint.cs
@@ -11,10 +11,11 @@ using osu.Framework.Graphics.Shapes;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
 using osu.Game.Rulesets.Edit;
+using osu.Game.Skinning;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Skinning.Editor
+namespace osu.Game.Overlays.SkinEditor
 {
     public partial class SkinBlueprint : SelectionBlueprint<ISkinnableDrawable>
     {
